+++ src/pages/[name]/[title]/index.tsx
@@ -6,20 +6,11 @@ import type {
 } from "next";
 import { useRouter } from "next/router";
 import useSWR from "swr";
-import useSWRInfinite from "swr/infinite";
 import Link from "next/link";
-import { useCallback, useRef, useState } from "react";
-import { useForm } from "react-hook-form";
-import { toast } from "react-toastify";
+import { useState } from "react";
 
 // type
-import {
-  Comment,
-  ICON,
-  Post as PostType,
-  SimplePost,
-  SimpleUser,
-} from "@src/types";
+import { ICON, Post as PostType, SimplePost } from "@src/types";
 
 // common-component
 import Spinner from "@src/components/common/Spinner";
@@ -27,11 +18,13 @@ import Photo from "@src/components/common/Photo";
 import Markdown from "@src/components/common/Markdown";
 import Post from "@src/components/Post";
 import Icon from "@src/components/common/Icon";
-import Button from "@src/components/common/Button";
 import Modal from "@src/components/common/Modal";
 
+// component
+import CommentContainer from "@src/components/Comment/CommentContainer";
+
 // util
-import { dateFormat, timeFormat } from "@src/libs/dateFormat";
+import { dateFormat } from "@src/libs/dateFormat";
 import { combineClassNames } from "@src/libs/util";
 
 // hook
@@ -54,16 +47,6 @@ type RelevantPostsResponse = {
   ok: boolean;
   posts: SimplePost[];
 };
-type CommentForm = {
-  comment: string;
-};
-interface ICommentWithUser extends Comment {
-  user: SimpleUser;
-}
-type CommentsResponse = {
-  ok: boolean;
-  comments: ICommentWithUser[];
-};
 type PostRemoveResponse = {
   ok: boolean;
 };
@@ -82,107 +65,6 @@ const PostDetail: NextPage<PostResponse> = ({ ok, post }) => {
   // 2022/04/30 - 현재 게시글과 연관된 게시글들 - by 1-blue
   const { data: relevantPosts } = useSWR<RelevantPostsResponse>(
     router.query.title ? `/api/post/${router.query.title}/relevant` : null
-  );
-
-  // 2022/04/30 - 댓글 입력 관련 메서드들 - by 1-blue
-  const { handleSubmit, register, reset } = useForm<CommentForm>();
-  // 2022/04/30 - comment Ref - by 1-blue
-  const { ref, ...rest } = register("comment");
-  const commentRef = useRef<HTMLTextAreaElement | null>(null);
-  // 2022/04/30 - textarea 자동 높이 조절 - by 1-blue
-  const handleResizeHeight = useCallback(() => {
-    if (!commentRef.current) return;
-
-    commentRef.current.style.height = "auto";
-    commentRef.current.style.height = commentRef.current?.scrollHeight + "px";
-  }, [commentRef]);
-
-  // 2022/05/02 - 댓글 추가 패치 가능 여부 - by 1-blue
-  const [hasMoreComment, setHasMoreComment] = useState(true);
-  // 2022/05/02 - 댓글들 순차적 요청 - by 1-blue
-  const {
-    data: commentsResponse,
-    setSize,
-    mutate: commentsMuate,
-    isValidating: commentsLoading,
-  } = useSWRInfinite<CommentsResponse>(
-    router.query.title
-      ? (pageIndex, previousPageData) => {
-          if (previousPageData && previousPageData.comments.length !== 10) {
-            setHasMoreComment(false);
-            return null;
-          }
-          if (previousPageData && !previousPageData.comments.length) {
-            setHasMoreComment(false);
-            return null;
-          }
-          return `/api/post/${
-            router.query.title
-          }/comment?page=${pageIndex}&offset=${10}`;
-        }
-      : () => null
-  );
-  // 2022/05/02 - 댓글 추가 관련 메서드 - by 1-blue
-  const [addComment, { loading: addCommentLoading }] = useMutation({
-    url: router.query.title ? `/api/post/${router.query.title}/comment` : null,
-    method: "POST",
-  });
-  // 2022/05/02 - 댓글 추가 - by 1-blue
-  const onAddComment = useCallback(
-    (body: CommentForm) => {
-      if (body.comment.length === 0)
-        return toast.error("댓글을 입력하고 제출해주세요!");
-      if (addCommentLoading)
-        return toast.error(
-          "댓글을 생성하는 중입니다.\n잠시후에 다시 시도해주세요!"
-        );
-
-      addComment({ contents: body.comment });
-
-      commentsMuate(
-        (prev) =>
-          prev && [
-            {
-              ok: true,
-              comments: [
-                {
-                  idx: Date.now(),
-                  contents: body.comment,
-                  postIdx: post.id,
-                  createdAt: new Date(Date.now()),
-                  updatedAt: new Date(Date.now()),
-                  user: me!,
-                  commentIdx: undefined,
-                },
-              ],
-            },
-            ...prev,
-          ],
-        false
-      );
-
-      reset();
-    },
-    [addCommentLoading, addComment, commentsMuate, post, me, reset]
-  );
-  // 2022/05/02 - 댓글 삭제 - by 1-blue
-  const onRemoveComment = useCallback(
-    (commentIdx: number) => async () => {
-      await fetch(`/api/post/${router.query.title}/comment/${commentIdx}`, {
-        method: "DELETE",
-      });
-
-      commentsMuate(
-        (prev) =>
-          prev &&
-          prev.map(({ comments }) => ({
-            ok: true,
-            comments: comments.filter((comment) => comment.idx !== commentIdx),
-          })),
-        false
-      );
-    },
-    [router, commentsMuate]
   );
 
   // 2022/05/01 - 게시글 삭제 모달 - by 1-blue
@@ -336,89 +218,8 @@ const PostDetail: NextPage<PostResponse> = ({ ok, post }) => {
 
         <hr />
 
-        {/* 댓글 작성 */}
-        <section className="space-y-4">
-          <span className="font-semibold">{"n"}개의 댓글</span>
-          <form className="space-y-4" onSubmit={handleSubmit(onAddComment)}>
-            <textarea
-              placeholder="댓글을 작성하세요"
-              {...rest}
-              className="w-full p-4 focus:outline-none resize-none rounded-sm bg-zinc-200 dark:bg-zinc-600"
-              onInput={handleResizeHeight}
-              ref={(e) => {
-                ref(e);
-                commentRef.current = e;
-              }}
-            />
-            <Button
-              type="submit"
-              className="block ml-auto font-semibold bg-indigo-400 text-white dark:bg-indigo-500 py-2 px-4 rounded-md"
-              contents="댓글 작성"
-              loading={addCommentLoading}
-            />
-          </form>
-        </section>
-
-        {/* 댓글들 */}
-        <section>
-          <>
-            {commentsResponse?.map(({ comments }, index) => (
-              <ul key={index} className="divide-y dark:divide-gray-400">
-                {comments.map((comment) => (
-                  <li key={comment.idx} className="space-y-4 pt-4">
-                    {/* 아바타, 이름, 작성시간, 삭제 버튼 */}
-                    <div className="flex space-x-2">
-                      <Photo
-                        photo={comment.user.avatar}
-                        size="w-14 h-14"
-                        alt="유저 이미지"
-                        $rouneded
-                      />
-                      <div className="flex flex-col">
-                        <span className="font-semibold">
-                          {comment.user.name}
-                        </span>
-                        <time className="text-sm dark:text-gray-400">
-                          {timeFormat(comment.updatedAt)}
-                        </time>
-                      </div>
-                      <div className="flex-1" />
-                      {comment.user.id === me?.id && (
-                        <button
-                          type="button"
-                          className="self-start text-gray-400 hover:text-white"
-                          onClick={onRemoveComment(comment.idx)}
-                        >
-                          삭제
-                        </button>
-                      )}
-                    </div>
-
-                    {/* 내용 */}
-                    <p className="whitespace-pre-line">{comment.contents}</p>
-
-                    {/* 답글 */}
-                    <div></div>
-                  </li>
-                ))}
-              </ul>
-            ))}
-
-            {hasMoreComment ? (
-              <Button
-                type="button"
-                className="block mx-auto px-4 py-2 rounded-md font-semibold text-white bg-indigo-400 hover:bg-indigo-500 dark:bg-indigo-500 dark:hover:bg-indigo-600"
-                contents="댓글 더 불러오기"
-                onClick={() => setSize((prev) => prev + 1)}
-                loading={commentsLoading}
-              />
-            ) : (
-              <span className="block text-center text-xl font-semibold">
-                더 이상 불러올 댓글이 없습니다.
-              </span>
-            )}
-          </>
-        </section>
+        {/* 댓글 영역 */}
+        <CommentContainer postIdx={post.id} />
       </article>
 
       <hr />
